Add a redirect for trivia night
We're blessed to have Jean Zeiler and her family organize our annual
trivia night. I want to have a record of each event so I've moved the
content to the "news" section and added a redirect at the old URL. We
can now add new pages for each year and change the redirect to point
to the new page.